Fix UI glitches when recording and following the playhead

+++ Source/UI/Sequencer/Header/Playhead.cpp
@@ -184,7 +184,7 @@ void Playhead::parentChanged()
 
 void Playhead::updatePosition(double position)
 {
-    const int newX = this->roll.getPlayheadPositionByBeat(position, float(this->getParentWidth()));
+    const int newX = this->roll.getPlayheadPositionByBeat(position, double(this->getParentWidth()));
     this->setTopLeftPosition(newX, 0);
 
     if (this->listener != nullptr)
